+++ src/main.ts
@@ -2,7 +2,7 @@ import helmet from 'helmet';
 import { NestFactory } from '@nestjs/core';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { AppModule } from './app.module';
-import { loggerMiddleware } from 'middlewares';
+import { loggerMiddleware } from './middlewares';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, { cors: true });
